L flag added

+++ srcs/double_utils.c
@@ -6,13 +6,13 @@
 /*   By: juligonz <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/25 17:36:20 by juligonz          #+#    #+#             */
-/*   Updated: 2019/11/25 17:36:21 by juligonz         ###   ########.fr       */
+/*   Updated: 2019/11/25 19:26:27 by juligonz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static void	put_exp(int64_t exp, t_manager *p)
+static void		put_exp(int64_t exp, t_manager *p)
 {
 	char c;
 
@@ -22,7 +22,7 @@ static void	put_exp(int64_t exp, t_manager *p)
 	write_buffer(p, &c, 1);
 }
 
-static void	put_mantis(double mantis, int precision, t_manager *p)
+static void		put_mantis(long double mantis, int precision, t_manager *p)
 {
 	char c;
 
@@ -33,9 +33,9 @@ static void	put_mantis(double mantis, int precision, t_manager *p)
 		put_mantis(mantis - (int)mantis, --precision, p);
 }
 
-double		ft_round(double n, size_t precision)
+long double		ft_round(long double n, size_t precision)
 {
-	double	rnd;
+	long double	rnd;
 
 	rnd = 0.5;
 	while (precision-- > 0)
@@ -43,7 +43,7 @@ double		ft_round(double n, size_t precision)
 	return (n + rnd);
 }
 
-void		put_double(double n, t_manager *p)
+void			put_double(long double n, t_manager *p)
 {
 	put_exp((int64_t)n, p);
 	if (p->precision > 0 || GET(F_HASH))
